pyproject.toml
[tool.poetry]
name = "higacoffeedbback"
version = "0.1.0"
description = ""
authors = ["Kaihotsu Kotaro <kaihotsu.kotaro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
SQLAlchemy = "^1.3.23"
Flask-JWT-Extended = "^4.1.0"
Flask-SQLAlchemy = "^2.4.4"
Flask-Cors = "^3.0.10"
bcrypt = "^3.2.0"
Flask-Bcrypt = "^0.7.1"
python-dotenv = "^0.15.0"
pytz = "^2021.1"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
flake8 = "^3.9.0"
pylint = "^2.7.2"
Flask-Testing = "^0.8.1"
yapf = "^0.31.0"
Flask-Migrate = "^2.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
